Refactor create methods of app into module
- Add rentals to attr_accessor
- integrated welcome commands into main methods
